[PhpUnitBridge] Added support for PHPUnit 7 in Coverage Listener
