+++ src/Symfony/Bridge/PhpUnit/Tests/CoverageListenerTest.php
@@ -16,7 +16,9 @@ class CoverageListenerTest extends TestCase
             $this->markTestSkipped('This test cannot be run on HHVM.');
         }
 
-        if (\PHP_VERSION_ID >= 70000) {
+        exec('type phpdbg', $output, $returnCode);
+
+        if (\PHP_VERSION_ID >= 70000 && 0 === $returnCode) {
             $php = 'phpdbg -qrr';
         } else {
             exec('php --ri xdebug -d zend_extension=xdebug.so 2> /dev/null', $output, $returnCode);
